backend/requirements.txt
langchain
langchain-community
huggingface_hub
python-dotenv
chromadb
PyPDF2
fastapi
uvicorn
sentence-transformers
torch
